Refactor ButterworthFilter and child classes to dataclasss
All classes are now keyword only dataclasses. Keyword only helps with
the fact that filter_type is expected for ButterworthFilter, but not
for child classes. The order of the arguments was fixed by
ButterworthFilter. Setting filter_type as the last argument did not make
sense.

The SpecifiedButterworthFilter intermediate class is no longer necessary
and is removed.

This commit also now allows the cutoff frequency to be any sequence when
using the band pass or band stop filters.