Introduce SelectByIndex and SelectByTime mixins
SelectByIndex and SelectByTime introduce the `select_by_index()` and
`select_by_time()` methods, as well as `__get_item__()` for slicing
by index, and the TimeIndexer `t` for slicing by time to an object.

Any object that inherits from from SelectByIndex and/or SelectByTime
has to implement the  method `_sliced_copy()` that returns a new
instance of the same class, but with a sliced version of the data.

These mixins replace object-specific implementations