Fix concatenation of data collections
Data collections could be None before, but not anymore. However, if a datacollection is empty, it equals False if cast to a boolean. If e.g. a.sparse_data and b.sparse_data were both empty, concat_sparse would be None. This would result in a failed equality test, because type(None) != type(DataCollection).